pyproject.toml
[project]
name = "fynautoserver"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuvraj",email = "yuvraj.r@kansoftware.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "beanie (>=1.30.0,<2.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "pymongo (>=4.13.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "fonttools (>=4.58.5,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jwt (>=4.1.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "fynautoserver", from = "src"}]

[tool.poetry.scripts]
dev = "fynautoserver.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
